# swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "AFSE API",
    "description": "API for Managing AFSE",
    "contact": {
      "name": "API Support",
      "email": "matteo.martinelli5@studenti.unimi.it"
    }
  },
  "host": "localhost:3000",
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local server"
    }
  ],
  "basePath": "/",
  "schemes": [
    "http"
  ],
  "paths": {
...
    "/shop/availablePacks": {
      "get": {
        "tags": [
          "Shop"
        ],
        "summary": "Recupera i pacchetti disponibili",
        "description": "Ritorna una lista di pacchetti disponibili per l acquisto.",
        "responses": {
          "200": {
            "description": "Lista dei pacchetti recuperata con successo",
            "schema": {
              "type": "object",
              "properties": {
                "type": {
                  "type": "string",
                  "example": "array"
                },
                "items": {
                  "type": "object",
                  "properties": {
                    "type": {
                      "type": "string",
                      "example": "object"
                    },
                    "properties": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "object",
                          "properties": {
                            "type": {
                              "type": "string",
                              "example": "integer"
                            },
                            "description": {
                              "type": "string",
                              "example": "ID del pacchetto"
                            },
                            "example": {
                              "type": "number",
                              "example": 1
                            }
                          }
                        },
                        "name": {
                          "type": "object",
                          "properties": {
                            "type": {
                              "type": "string",
                              "example": "string"
                            },
                            "description": {
                              "type": "string",
                              "example": "Nome del pacchetto"
                            },
                            "example": {
                              "type": "string",
                              "example": "Pacchetto Premium"
                            }
                          }
                        },
                        "price": {
                          "type": "object",
                          "properties": {
                            "type": {
                              "type": "string",
                              "example": "number"
                            },
                            "description": {
                              "type": "string",
                              "example": "Prezzo del pacchetto"
                            },
                            "example": {
                              "type": "number",
                              "example": 49.99
                            }
                          }
                        }
                      }
                    }
                  }
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "400": {
            "description": "Richiesta non valida (bad request)",
            "schema": {
              "type": "object",
              "properties": {
                "type": {
                  "type": "string",
                  "example": "object"
                },
                "properties": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "object",
                      "properties": {
                        "type": {
                          "type": "string",
                          "example": "string"
                        },
                        "example": {
                          "type": "string",
                          "example": "Bad request"
                        }
                      }
                    }
                  }
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Errore interno del server",
            "schema": {
              "type": "object",
              "properties": {
                "type": {
                  "type": "string",
                  "example": "object"
                },
                "properties": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "object",
                      "properties": {
                        "type": {
                          "type": "string",
                          "example": "string"
                        },
                        "example": {
                          "type": "string",
                          "example": "Internal server error"
                        }
                      }
                    }
                  }
                }
              },
              "xml": {
                "name": "main"
              }
            }
          }
        },
        "security": [
          {
            "cookieAuth": []
          }
        ]
      }
    },
    "/shop/coinsOffers": {
      "get": {
        "tags": [
          "Shop"
        ],
        "summary": "Recupera le offerte di monete",
        "description": "Ritorna una lista di offerte di monete disponibili per l acquisto.",
        "responses": {
          "200": {
            "description": "Lista delle offerte di monete recuperata con successo",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "id": {
                        "type": "integer",
                        "description": "ID dell offerta di monete",
                        "example": 1
                      },
                      "name": {
                        "type": "string",
                        "description": "Nome dell offerta",
                        "example": "Offerta base"
                      },
                      "coinAmount": {
                        "type": "integer",
                        "description": "Quantità di monete incluse nell offerta",
                        "example": 100
                      },
                      "price": {
                        "type": "number",
                        "description": "Prezzo dell offerta",
                        "example": 9.99
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Richiesta non valida (bad request)",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Bad request"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Errore interno del server",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Internal server error"
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "cookieAuth": []
          }
        ]
      }
    },
    "/shop/buyCoin": {
      "post": {
        "tags": [
          "Shop"
        ],
        "summary": "Compra un offerta di monete",
        "description": "Permette all utente di acquistare una determinata quantità di monete.",
        "parameters": [
          {
            "name": "obj",
            "in": "body",
            "description": "ID dell offerta di monete da acquistare (24 caratteri, ID MongoDB)",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "type": {
                  "type": "string",
                  "example": "object"
                },
                "properties": {
                  "type": "object",
                  "properties": {
                    "coinOfferId": {
                      "type": "object",
                      "properties": {
                        "type": {
                          "type": "string",
                          "example": "string"
                        },
                        "example": {
                          "type": "string",
                          "example": "507f1f77bcf86cd799439011"
                        }
                      }
                    }
                  }
                },
                "required": {
                  "type": "array",
                  "example": [
                    "coinOfferId"
                  ],
                  "items": {
                    "type": "string"
                  }
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Acquisto completato con successo",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "coinAmount": {
                      "type": "integer",
                      "description": "Quantità di monete acquistate",
                      "example": 100
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Errore nella richiesta o errore interno durante la transazione",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "description": "Messaggio di errore",
                      "example": "Cannot buy coins for internal error"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Errore interno del server",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "description": "Messaggio di errore",
                      "example": "Internal server error"
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "cookieAuth": []
          }
        ]
      }
    },
    "/shop/buyPack": {
      "post": {
        "tags": [
          "Shop"
        ],
        "summary": "Compra un pacchetto di carte",
        "description": "Permette all\\'utente di acquistare un pacchetto di carte se ha abbastanza monete. Riduce la quantità di monete e restituisce l\\'ID del pacchetto acquistato.",
        "parameters": [
          {
            "name": "obj",
            "in": "body",
            "description": "ID del pacchetto di carte da acquistare (24 caratteri, ID MongoDB)",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "type": {
                  "type": "string",
                  "example": "object"
                },
                "properties": {
                  "type": "object",
                  "properties": {
                    "packId": {
                      "type": "object",
                      "properties": {
                        "type": {
                          "type": "string",
                          "example": "string"
                        },
                        "description": {
                          "type": "string",
                          "example": "ID del pacchetto di carte da acquistare"
                        },
                        "example": {
                          "type": "string",
                          "example": "507f1f77bcf86cd799439011"
                        }
                      }
                    }
                  }
                },
                "required": {
                  "type": "array",
                  "example": [
                    "packId"
                  ],
                  "items": {
                    "type": "string"
                  }
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Acquisto del pacchetto completato con successo",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "packId": {
                      "type": "string",
                      "description": "ID del pacchetto acquistato",
                      "example": "60d5f95f0b8f4c5b8f8d77d3"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Errore nella richiesta o fondi insufficienti",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "description": "Messaggio di errore",
                      "example": "Insufficient funds"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Errore interno del server",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "description": "Messaggio di errore",
                      "example": "Internal error"
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "cookieAuth": []
          }
        ]
      }
    },
...
  },
...
  "components": {
    "securitySchemes": {
      "cookieAuth": {
        "type": "jwt",
        "in": "cookie",
        "name": "jwt",
        "description": "JWT token passato tramite cookie"
      }
    },
...
  }
}